\name{mztwinreg-package}
\alias{mztwinreg-package}
\alias{mztwinreg}
\docType{package}
\title{
Regression Modeling of MZ Twin Data 
}
\description{
Linear and logistic regression models for quantitative genetic analysis of data from monozygotic twins
}
\details{
\tabular{ll}{
Package: \tab mztwinreg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-13\cr
License: \tab What license is it under?\cr
}
This package implements a number of functions to perform different linear and logistic regression models of data from monozygotic twins. Examples include the variability gene approach by Berg, K. (1994), the Epigenetic Epidemilogy model by Tan, Q. (2013)., and the conventional approaches of Carlin, J. B. et al. (1994). 
}
\author{
Aldo Cordova-Palomera

Maintainer: Aldo Cordova-Palomera <a.cordova@ub.edu>
}
\references{
Berg, K. (1994). Gene-environment interaction: variability gene concept. In Genetic factors in coronary heart disease (pp. 373-383). Springer Netherlands.

Carlin, J. B., Gurrin, L. C., Sterne, J. A., Morley, R., & Dwyer, T. (2005). Regression models for twin studies: a critical review. International Journal of Epidemiology, 34(5), 1089-1099.

Cordova-Palomera, A., Alemany, S., Fatjo-Vilas, M., Goldberg, X., Leza, J. C., Gonzalez-Pinto, A., Nenadic, I., & Fananas, L. (2014). Birth Weight, Working Memory and Epigenetic Signatures in IGF2 and Related Genes: A MZ Twin Study. PloS one, 9(8), e103639.

Cordova-Palomera, A., Fatjo-Vilas, M., Kebir, O., Gasto, C., Krebs, M. O., & Fananas, L. (2014). Polymorphic variation in the epigenetic gene DNMT3B modulates the environmental impact on cognitive ability: A twin study. European Psychiatry.

Tan, Q. (2013). Epigenetic Epidemiology of Complex Diseases Using Twins. Medical Epigenetics, 1(1), 46-51.

Tan, Q., Frost, M., Heijmans, B. T., von Bornemann Hjelmborg, J., Tobi, E. W., Christensen, K., & Christiansen, L. (2014). Epigenetic signature of birth weight discordance in adult twins. BMC genomics, 15(1), 1062.

}
